testapp/requirements.txt
streamlit==1.32.0
Pillow==10.2.0
numpy
reportlab
streamlit-drawable-canvas==0.9.3
pandas==2.2.0
python-docx==1.1.0
openpyxl==3.1.2
img2pdf==0.5.1
holidays==0.38
qrcode==7.4.2
python-dateutil